pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "clients",
    "config",
    "data",
    "engine",
    "execution",
    "interfaces",
    "portfolio",
    "reporting",
    "risk",
    "utils",
    "workers",
]

[project]
name = "infrastructure"
version = "0.1.0"
description = "Core infrastructure of the fund"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "lighter-sdk",
    "nats-py>=2.6.0",
    "websockets>=16.0",
]  # runtime deps go here as modules are built out

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "pandas-stubs",
    "types-PyYAML",
    "types-requests",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pip-audit>=2.7",
    "pre-commit>=4.5.1",
    "detect-secrets>=1.5.0",
]

# [tool.mypy] — config lives here, not a separate mypy.ini
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
warn_unused_ignores = true
warn_return_any = true

# [tool.ruff] — replaces your need for a separate .ruff.toml
[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
# B = bugbear, SIM = simplify, TCH = type-checking imports

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests

# [tool.bandit] — your pre-commit hook references pyproject.toml for bandit config
[tool.bandit]
skips = ["B101"]        # assert statements — fine in tests
exclude_dirs = ["tests", "docs"]
severity = "high"       # only high severity, as you intended

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=75"
markers = [
    "unit: fast, no I/O",
    "integration: requires external mocks",
    "regression: numerical output pinning",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.uv.sources]
lighter-sdk = { git = "https://github.com/elliottech/lighter-python.git" }
